refactor(modal): remove ModalPortal in favor of Modal
