+++ packages/kit-headless/src/components/modal/index.ts
@@ -2,5 +2,4 @@ export * from './modal';
 export * from './modal-content';
 export * from './modal-footer';
 export * from './modal-header';
-export * from './modal-popup';
 export * from './types';
