+++ packages/kit-headless/src/components/modal/modal.tsx
@@ -1,16 +1,21 @@
 import {
+  $,
   component$,
   QRL,
+  QwikIntrinsicElements,
+  QwikMouseEvent,
   Signal,
   Slot,
   useContextProvider,
   useSignal,
   useTask$,
+  useVisibleTask$,
 } from '@builder.io/qwik';
+import { createFocusTrap, FocusTrap } from 'focus-trap';
 import { modalContextId } from './modal-context-id';
 import { ModalContext } from './modal-context.type';
 
-export type ModalProps = {
+export type ModalProps = Omit<QwikIntrinsicElements['dialog'], 'open'> & {
   onShow$?: QRL<() => void>;
   onHide$?: QRL<() => void>;
   show?: boolean;
@@ -18,8 +23,9 @@ export type ModalProps = {
 };
 
 export const Modal = component$((props: ModalProps) => {
-  const { 'bind:show': givenOpenSig, show: givenShow, onShow$, onHide$ } = props;
+  const refSig = useSignal<HTMLDialogElement>();
 
+  const { 'bind:show': givenOpenSig, show: givenShow, onShow$, onHide$ } = props;
   const defaultOpenSig = useSignal(false);
   const showSig = givenOpenSig || defaultOpenSig;
 
@@ -29,6 +35,77 @@ export const Modal = component$((props: ModalProps) => {
     showSig.value = openPropValue || false;
   });
 
+  useTask$(async function openOrCloseModal({ track }) {
+    const isOpen = track(() => showSig.value);
+
+    const modal = refSig.value;
+
+    if (!modal) return;
+
+    if (isOpen) {
+      modal.showModal();
+      await props.onShow$?.();
+
+      let scrollbarWidth: number | null = null;
+
+      // prevents scrollbar flickers
+      if (scrollbarWidth === null) {
+        scrollbarWidth = window.innerWidth - document.body.offsetWidth;
+        document.body.style.overflow = 'hidden';
+        document.body.style.paddingRight = `${scrollbarWidth}px`;
+      }
+    } else {
+      await props.onHide$?.();
+      modal.close();
+    }
+  });
+
+  useVisibleTask$(function toggleFocusTrap({ track, cleanup }) {
+    const isOpen = track(() => showSig.value);
+    const modal = refSig.value;
+
+    if (!modal) return;
+
+    let focusTrap: FocusTrap | null = createFocusTrap(modal, {
+      escapeDeactivates: false,
+    });
+
+    if (isOpen) {
+      focusTrap.activate();
+    }
+
+    cleanup(() => {
+      focusTrap?.deactivate();
+      focusTrap = null;
+    });
+  });
+
+  useVisibleTask$(function lockScrollingWhenModalIsOpen({ track }) {
+    const isOpen = track(() => showSig.value);
+
+    window.document.body.style.overflow = isOpen ? 'hidden' : '';
+  });
+
+  const closeOnBackdropClick$ = $((event: QwikMouseEvent) => {
+    const modal = refSig.value;
+
+    if (!modal) {
+      return;
+    }
+
+    const modalRect = modal.getBoundingClientRect();
+
+    const wasClickTriggeredOutsideModalRect =
+      modalRect.left > event.clientX ||
+      modalRect.right < event.clientX ||
+      modalRect.top > event.clientY ||
+      modalRect.bottom < event.clientY;
+
+    if (wasClickTriggeredOutsideModalRect) {
+      showSig.value = false;
+    }
+  });
+
   const context: ModalContext = {
     showSig,
     handler: { onShow$, onHide$ },
@@ -36,5 +113,14 @@ export const Modal = component$((props: ModalProps) => {
 
   useContextProvider(modalContextId, context);
 
-  return <Slot />;
+  return (
+    <dialog
+      {...props}
+      ref={refSig}
+      onClick$={(event) => closeOnBackdropClick$(event)}
+      onClose$={() => (showSig.value = false)}
+    >
+      <Slot />
+    </dialog>
+  );
 });
